requirements.txt
build==0.10.0
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
docutils==0.19 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
imagesize==1.4.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
iniconfig==2.0.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
nose==1.3.7
packaging==23.0 ; python_version >= '3.7'
path==16.6.0 ; python_version >= '3.7'
path.py==12.5.0
pathlib==1.0.1
pluggy==1.0.0 ; python_version >= '3.6'
pygments==2.14.0 ; python_version >= '3.6'
pyproject-hooks==1.0.0 ; python_version >= '3.7'
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
